main/requirements.txt
django==4.0
paho-mqtt==1.5.0
pynput==1.7.6
six==1.16.0
xerox==0.4.1